v2.1.0-rc+manual-20251208.9654dfb
* v2.1.0-rc+manual-20250918.1335b03

* v2.1.0-rc+manual-20250929.026a60b

* v2.1.0-rc+manual-20250929.967b9cb

* v2.1.0-rc+manual-20251024.9f51199

* v2.1.0-rc+manual-20251029.09fcbf5

* v2.1.0-rc+manual-20251112.8bbd962

* v2.1.0-rc+manual-20251118.6eb14c7

* v2.1.0-rc+manual-20251119.d1e5783

* v2.1.0-rc+manual-20251125.27baf1b

* v2.1.0-rc+manual-20251126.a029667

* v2.1.0-rc+manual-20251128.6b558b5

* v2.1.0-rc+manual-20251128.b622589

* v2.1.0-rc+manual-20251202.bf0dc56

* v2.1.0-rc+manual-20251203.e44fff2

* Update DX-APP release note

Update DX-APP release note

* v2.1.0-rc+manual-20251208.9654dfb

---------

+++ demos/demo_utils/segmentation.cpp
@@ -26,9 +26,9 @@ void Segmentation(float *input, uint8_t *output, int rows, int cols, Segmentatio
      *  When need_transpose is True (use-ort on mode), the output data format is [1, 19, 640, 640].
      *  Test results show that FPS improves by 1.5x when use-ort is off.
      */
-    bool need_transpose = shape[1] == numClasses? true : false;
+    bool need_transpose = static_cast<int>(shape[1]) == numClasses? true : false;
     int compare_max_idx, compare_channel_idx;
-    int pitch = shape[3];
+    int pitch = static_cast<int>(shape[3]);
     for(int h=0;h<rows;h++)
     {
         for(int w=0;w<cols;w++)
